[update]超音波センサによる距離の取得をint型へ

+++ share/PerceptionDataAccess.h
@@ -15,7 +15,7 @@ struct PerceptionDataAccess
     /**
      * 超音波センサーの値
      */
-    float distance;
+    int distance;
 
     /**
      * カラーセンサーの値
